requirements.txt
pandas
pandarallel
python-chess
scikit-learn
numpy
joblib
tqdm
pycaret
xgboost